requirement.txt
Django==4.2.2
asgiref
Pillow
sqlparse
typing_extensions
tzdata
django_restframework